correcciones papeletas
