+++ package.json
@@ -54,7 +54,7 @@
     "lucide-react": "^0.414.0",
     "moment-timezone": "^0.5.47",
     "multer": "^1.4.5-lts.1",
-    "mysql2": "^3.11.5",
+    "mysql2": "^3.12.0",
     "next": "^14.2.20",
     "next-auth": "^4.24.7",
     "nodemailer": "^6.9.15",
@@ -69,6 +69,7 @@
     "react-tooltip": "^4.2.21",
     "recharts": "^2.12.7",
     "redis": "^4.7.0",
+    "sequelize": "^6.37.5",
     "ssh2": "^1.15.0",
     "sweetalert2": "^11.14.0",
     "sweetalert2-react-content": "^5.0.7",
